+++ internal/helper/fieldextractors/repository.go
@@ -1,6 +1,10 @@
 package fieldextractors
 
-import pb "gitlab.com/gitlab-org/gitaly-proto/go"
+import (
+	"strings"
+
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+)
 
 func formatRepoRequest(repo *pb.Repository) map[string]interface{} {
 	if repo == nil {
@@ -12,9 +16,24 @@ func formatRepoRequest(repo *pb.Repository) map[string]interface{} {
 	}
 
 	return map[string]interface{}{
-		"repoStorage": repo.StorageName,
-		"repoPath":    repo.RelativePath,
+		"repoStorage":   repo.StorageName,
+		"repoPath":      repo.RelativePath,
+		"topLevelGroup": getTopLevelGroupFromRepoPath(repo.RelativePath),
 	}
+}
+
+// getTopLevelGroupFromRepoPath gives the top-level group name, given
+// a repoPath. For example:
+// - "gitlab-org/gitlab-ce.git" returns "gitlab-org"
+// - "gitlab-org/gitter/webapp.git" returns "gitlab-org"
+// - "x.git" returns ""
+func getTopLevelGroupFromRepoPath(repoPath string) string {
+	parts := strings.SplitN(repoPath, "/", 2)
+	if len(parts) != 2 {
+		return ""
+	}
+
+	return parts[0]
 }
 
 type repositoryBasedRequest interface {